+++ src/main/java/org/bytedeco/javacpp/tools/TemplateMap.java
@@ -23,6 +23,7 @@
 package org.bytedeco.javacpp.tools;
 
 import java.util.LinkedHashMap;
+import java.util.Map;
 
 /**
  *
@@ -66,5 +67,21 @@ class TemplateMap extends LinkedHashMap<String,Type> {
         } else {
             return value;
         }
+    }
+
+    @Override
+    public String toString() {
+        String s = "<";
+        for (Map.Entry<String, Type> e : entrySet()) {
+            if (s.length() > 1) {
+                s += ",";
+            }
+            Type t = e.getValue();
+            s += t != null ? t.cppName : e.getKey();
+        }
+        if (s.endsWith(">")) {
+            s += " ";
+        }
+        return s + ">";
     }
 }