+++ src/main/java/org/bytedeco/javacpp/tools/Templates.java
@@ -47,8 +47,17 @@ class Templates {
         return strip(s).length() == s.length();
     }
 
-    /** Split s at ::, but taking care of qualified template arguments */
+    /** Returns {@code splitNamespace(s, false)}. */
     static List<String> splitNamespace(String s) {
+        return splitNamespace(s, false);
+    }
+
+    /**
+     * Split s at ::, but taking care of qualified template arguments and qualified function parameters.
+     * If returnParams is true, returned list contains an extra element with function parameters, or the empty
+     * string if none are present.
+     */
+    static List<String> splitNamespace(String s, boolean returnParams) {
         String sTemplatesMasked = s;
         for (;;) {
             Matcher m = templatePattern.matcher(sTemplatesMasked);
@@ -61,17 +70,43 @@ class Templates {
             }
         }
         ArrayList<String> comps = new ArrayList<>();
+        int pIndex = sTemplatesMasked.lastIndexOf(')'); // last because of function pointer types like void(*)()
+        if (pIndex > 0) {
+            // Pointer list may contain function pointer types with parentheses
+            int count = 1;
+            for (pIndex--; pIndex >= 0; pIndex--) {
+                char c = sTemplatesMasked.charAt(pIndex);
+                if (c == ')') {
+                    count++;
+                } else if (c == '(') {
+                    count--;
+                    if (count == 0) {
+                        break;
+                    }
+                }
+            }
+        }
         int start = 0;
         for (;;) {
             int i = sTemplatesMasked.indexOf("::", start);
-            if (i >= 0) {
+            if (i >= 0 && (i < pIndex || pIndex == -1)) {
                 comps.add(s.substring(start, i));
                 start = i + 2;
             } else {
                 break;
             }
         }
-        comps.add(s.substring(start));
+        String params;
+        if (pIndex >= 0) {
+            comps.add(s.substring(start, pIndex));
+            params = s.substring(pIndex);
+        } else {
+            comps.add(s.substring(start));
+            params = "";
+        }
+        if (returnParams) {
+            comps.add(params);
+        }
         return comps;
     }
 }